+++ src/model/Invite.zig
@@ -4,7 +4,9 @@ const jconfig = @import("../root.zig").jconfig;
 const Omittable = jconfig.Omittable;
 const Partial = jconfig.Partial;
 
-// partial version of this object is included in `rest/endpoints/guild.zig` (GetGuildVanityUrlResponse), changes here should be reflected in changes there.
+const Invite = @This();
+
+// partial version of this object is included in `rest/endpoints/channel.zig`, changes here should be reflected in there.
 
 type: Type,
 code: []const u8,
@@ -19,6 +21,8 @@ approximate_member_count: Omittable(i64) = .omit,
 expires_at: Omittable(?[]const u8) = .omit,
 stage_instance: Omittable(InviteStageInstance) = .omit, // deprecated
 guild_scheduled_event: Omittable(model.GuildScheduledEvent) = .omit,
+flags: Omittable(Flags) = .omit,
+roles: Omittable(PartialRole) = .omit,
 
 pub const jsonStringify = jconfig.stringifyWithOmit;
 
@@ -38,19 +42,7 @@ pub const InviteStageInstance = struct {
 };
 
 pub const WithMetadata = struct {
-    type: Type,
-    code: []const u8,
-    guild: Omittable(model.guild.Guild) = .omit,
-    channel: ?model.Channel,
-    inviter: Omittable(model.User) = .omit,
-    target_type: Omittable(i64) = .omit,
-    target_user: Omittable(model.User) = .omit,
-    target_application: Omittable(model.Application) = .omit,
-    approximate_presence_count: Omittable(i64) = .omit,
-    approximate_member_count: Omittable(i64) = .omit,
-    expires_at: Omittable(?[]const u8) = .omit,
-    stage_instance: Omittable(InviteStageInstance) = .omit, // deprecated
-    guild_scheduled_event: Omittable(model.GuildScheduledEvent) = .omit,
+    invite: Invite,
 
     // extra metadata fields provided by some endpoints
     uses: i64,
@@ -58,6 +50,28 @@ pub const WithMetadata = struct {
     max_age: i64,
     temporary: bool,
     created_at: []const u8,
+
+    const Mixin = jconfig.InlineSingleStructFieldMixin(@This(), "invite");
+    pub const jsonStringify = Mixin.jsonStringify;
+    pub const jsonParse = Mixin.jsonParse;
+    pub const jsonParseFromValue = Mixin.jsonParseFromValue;
+};
+
+pub const Flags = packed struct(u64) {
+    is_guest_invite: bool = false,
+    _padding: u63 = 0,
+
+    const Mixin = model.PackedFlagsMixin(@This());
+};
+
+pub const PartialRole = struct {
+    id: model.Snowflake,
+    name: []const u8,
+    position: i64,
+    color: u64,
+    colors: model.Role.Colors,
+    icon: jconfig.Omittable(?[]const u8) = .omit,
+    unicode_emoji: jconfig.Omittable(?[]const u8) = .omit,
 
     pub const jsonStringify = jconfig.stringifyWithOmit;
 };